+++ src/main/kotlin/com/jetbrains/snakecharm/codeInsight/SnakemakeAPI.kt
@@ -5,9 +5,11 @@ import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_BENCHMARK
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_CACHE
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_CONDA
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_CONTAINER
+import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_CONTAINERIZED
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_CWL
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_ENVMODULES
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_GROUP
+import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_HANDOVER
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_INPUT
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_LOG
 import com.jetbrains.snakecharm.lang.SnakemakeNames.SECTION_MESSAGE
@@ -85,7 +87,8 @@ object SnakemakeAPI {
             SECTION_CWL, SECTION_BENCHMARK, SECTION_VERSION,
             SECTION_MESSAGE, SECTION_THREADS, SECTION_SINGULARITY,
             SECTION_PRIORITY, SECTION_CONDA, SECTION_GROUP,
-            SECTION_SHADOW, SECTION_CACHE, SECTION_NOTEBOOK, SECTION_CONTAINER
+            SECTION_SHADOW, SECTION_CACHE, SECTION_NOTEBOOK, SECTION_CONTAINER,
+            SECTION_HANDOVER, SECTION_CONTAINERIZED
     )
 
     /**
@@ -99,8 +102,10 @@ object SnakemakeAPI {
             SECTION_SCRIPT, SECTION_WRAPPER, SECTION_CWL, SECTION_NOTEBOOK,
             SECTION_CACHE,
             SECTION_CONTAINER,
+            SECTION_CONTAINERIZED,
             SECTION_ENVMODULES,
-            SECTION_NAME
+            SECTION_NAME,
+            SECTION_HANDOVER
     )
     val RULE_OR_CHECKPOINT_SECTION_KEYWORDS = (RULE_OR_CHECKPOINT_ARGS_SECTION_KEYWORDS + setOf(SnakemakeNames.SECTION_RUN))
 
@@ -152,7 +157,8 @@ object SnakemakeAPI {
             SECTION_WRAPPER,
             SECTION_VERSION, SECTION_THREADS,
             SECTION_PRIORITY, SECTION_SINGULARITY, SECTION_CACHE,
-            SECTION_CONTAINER, SECTION_NOTEBOOK, SECTION_ENVMODULES
+            SECTION_CONTAINER, SECTION_CONTAINERIZED, SECTION_NOTEBOOK,
+            SECTION_ENVMODULES, SECTION_HANDOVER
     )
 
     /**
@@ -210,7 +216,8 @@ object SnakemakeAPI {
     val SECTIONS_WHERE_KEYWORD_ARGS_PROHIBITED = setOf(
             SECTION_BENCHMARK, SECTION_VERSION, SECTION_MESSAGE, SECTION_SHELL, SECTION_THREADS, SECTION_SINGULARITY,
             SECTION_PRIORITY, SECTION_GROUP, SECTION_SHADOW, SECTION_CONDA, SECTION_SCRIPT, SECTION_WRAPPER,
-            SECTION_CWL, SECTION_NOTEBOOK, SECTION_CACHE, SECTION_CONTAINER, SECTION_ENVMODULES, SECTION_NAME
+            SECTION_CWL, SECTION_NOTEBOOK, SECTION_CACHE, SECTION_CONTAINER, SECTION_CONTAINERIZED, SECTION_ENVMODULES,
+            SECTION_NAME, SECTION_HANDOVER
     )
 
     val IO_FLAG_2_SUPPORTED_SECTION: HashMap<String, List<String>> = hashMapOf(