+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/SmkUnrecognizedSectionInspection.kt
@@ -1,14 +1,21 @@
 package com.jetbrains.snakecharm.inspections
 
 import com.intellij.codeInspection.LocalInspectionToolSession
+import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
+import com.intellij.codeInspection.ui.ListEditForm
 import com.jetbrains.snakecharm.SnakemakeBundle
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.RULE_OR_CHECKPOINT_ARGS_SECTION_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.SUBWORKFLOW_SECTIONS_KEYWORDS
+import com.jetbrains.snakecharm.inspections.quickfix.AddIgnoredElementQuickFix
 import com.jetbrains.snakecharm.lang.psi.SmkRuleOrCheckpointArgsSection
 import com.jetbrains.snakecharm.lang.psi.SmkSubworkflowArgsSection
+import javax.swing.JComponent
 
 class SmkUnrecognizedSectionInspection : SnakemakeInspection() {
+    @JvmField
+    val ignoredItems: ArrayList<String> = arrayListOf()
+
     override fun buildVisitor(
         holder: ProblemsHolder,
         isOnTheFly: Boolean,
@@ -18,10 +25,13 @@ class SmkUnrecognizedSectionInspection : SnakemakeInspection() {
             val sectionNamePsi = st.nameIdentifier
             val sectionKeyword = st.sectionKeyword
 
-            if (sectionNamePsi != null && sectionKeyword != null && sectionKeyword !in SUBWORKFLOW_SECTIONS_KEYWORDS) {
+            if (sectionNamePsi != null && sectionKeyword != null && sectionKeyword !in SUBWORKFLOW_SECTIONS_KEYWORDS
+                && sectionKeyword !in ignoredItems) {
                 registerProblem(
                     sectionNamePsi,
-                    SnakemakeBundle.message("INSP.NAME.section.unrecognized.message", sectionKeyword)
+                    SnakemakeBundle.message("INSP.NAME.section.unrecognized.message", sectionKeyword),
+                    ProblemHighlightType.WEAK_WARNING,
+                    null, AddIgnoredElementQuickFix(sectionKeyword)
                 )
             }
         }
@@ -30,10 +40,13 @@ class SmkUnrecognizedSectionInspection : SnakemakeInspection() {
             val sectionNamePsi = st.nameIdentifier
             val sectionKeyword = st.sectionKeyword
 
-            if (sectionNamePsi != null && sectionKeyword != null && sectionKeyword !in RULE_OR_CHECKPOINT_ARGS_SECTION_KEYWORDS) {
+            if (sectionNamePsi != null && sectionKeyword != null && sectionKeyword !in RULE_OR_CHECKPOINT_ARGS_SECTION_KEYWORDS
+                && sectionKeyword !in ignoredItems) {
                 registerProblem(
                     sectionNamePsi,
-                    SnakemakeBundle.message("INSP.NAME.section.unrecognized.message", sectionKeyword)
+                    SnakemakeBundle.message("INSP.NAME.section.unrecognized.message", sectionKeyword),
+                    ProblemHighlightType.WEAK_WARNING,
+                    null, AddIgnoredElementQuickFix(sectionKeyword)
                 )
             }
         }
@@ -44,4 +57,8 @@ class SmkUnrecognizedSectionInspection : SnakemakeInspection() {
         //     super.visitSmkWorkflowArgsSection(st)
         // }
     }
+
+    override fun createOptionsPanel(): JComponent? =
+        ListEditForm(SnakemakeBundle.message("INSP.NAME.section.unrecognized.ignored"), ignoredItems).contentPanel
 }
+