+++ src/main/kotlin/com/jetbrains/snakecharm/lang/SnakemakeNames.kt
@@ -30,6 +30,7 @@ object SnakemakeNames {
 
     const val WORKFLOW_ENVVARS_KEYWORD = "envvars"
     const val WORKFLOW_CONTAINER_KEYWORD = "container"
+    const val WORKFLOW_CONTAINERIZED_KEYWORD = "containerized"
 
     const val SECTION_INPUT = "input"
     const val SECTION_OUTPUT = "output"
@@ -53,9 +54,11 @@ object SnakemakeNames {
     const val SECTION_RUN = "run"
     const val SECTION_CACHE = "cache" // >= 5.12.0
     const val SECTION_CONTAINER = "container"
+    const val SECTION_CONTAINERIZED = "containerized" // => 6.0.0
     const val SECTION_NOTEBOOK = "notebook"
     const val SECTION_ENVMODULES = "envmodules" // >= 5.9
     const val SECTION_NAME = "name" // >= 5.31
+    const val SECTION_HANDOVER = "handover" // >= 6.2
 
     const val RUN_SECTION_VARIABLE_RULE = "rule"
     const val RUN_SECTION_VARIABLE_JOBID = "jobid"
@@ -70,4 +73,6 @@ object SnakemakeNames {
     const val SNAKEMAKE_IO_METHOD_REPEAT = "repeat"
     const val SNAKEMAKE_IO_METHOD_UNPACK = "unpack"
     const val SNAKEMAKE_IO_METHOD_DYNAMIC = "dynamic"
+
+    const val SNAKEMAKE_METHOD_MULTIEXT = "multiext"
 }