+++ CHANGELOG.md
@@ -7,10 +7,18 @@
 Released on ...
 
 ### Changed
+- Show argument preview for int indexes completion in `output[i]` (see [#378](https://github.com/JetBrains-Research/snakecharm/issues/378))
 - TODO  (see [#NNN](https://github.com/JetBrains-Research/snakecharm/issues/NNN)) 
 
 ### Fixed
 - Default path "config/config.yaml" now supported (see [#363](https://github.com/JetBrains-Research/snakecharm/issues/363))
+- Access by index to input/output sections with 'multiext' function (see [#278](https://github.com/JetBrains-Research/snakecharm/issues/278))
+- TODO  (see [#NNN](https://github.com/JetBrains-Research/snakecharm/issues/NNN))
+
+### Added
+- Ability for memorising new section name (see [#372](https://github.com/JetBrains-Research/snakecharm/issues/372))
+- Support for 'handover' section (see [#362](https://github.com/JetBrains-Research/snakecharm/issues/362))
+- Support for 'containerized' section (see [#361](https://github.com/JetBrains-Research/snakecharm/issues/361))
 - TODO  (see [#NNN](https://github.com/JetBrains-Research/snakecharm/issues/NNN))
 
 ## [2021.2.424]